Refactor: #34 - 카메라에 있는 함수를 extension으로 이동
