+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Camera/Camera+.swift
@@ -1,0 +1,38 @@
+//
+//  Camera+.swift
+//  2024-MacC-M4-6princess
+//
+//  Created by ram on 11/11/24.
+//
+
+import Foundation
+import SwiftUI
+import AVFoundation
+import CoreData
+
+extension CameraView {
+    //이미지 렌더링해서 불러오기
+    func loadSelectedFrame() {
+        guard let frameId = viewModel.selectedFrame else {
+            viewModel.frameImage = nil
+            return
+        }
+        
+        let fetchRequest: NSFetchRequest<StoreImages> = StoreImages.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "uuid == %@", frameId as CVarArg)
+        fetchRequest.fetchLimit = 1
+        
+        do {
+            let results = try viewContext.fetch(fetchRequest)
+            if let storedImage = results.first, let imageData = storedImage.image {
+                viewModel.frameImage = UIImage(data: imageData)
+                //                frameImage = UIImage(data: imageData)
+            } else {
+                viewModel.frameImage = nil
+            }
+        } catch {
+            print("Error fetching frame: \(error)")
+            viewModel.frameImage = nil
+        }
+    }
+}
