+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/Frame/ImageEdit/IEViewBuilder.swift
@@ -96,7 +96,7 @@ struct CustomSliderView: View {
             .padding(.horizontal)
         }
         .frame(width: viewModel.screenSize.width,height: 40)
-        //        .background(Color.black.opacity(0.5)) // 배경색
+                .background(Color.black.opacity(0.5)) // 배경색
     }
 }
